Move ClickType, InventoryAction to own files
