+++ src/main/java/org/bukkit/event/inventory/InventoryAction.java
@@ -1,0 +1,121 @@
+package org.bukkit.event.inventory;
+
+/**
+ * An estimation of what the result will be.
+ */
+public enum InventoryAction {
+    /**
+     * Nothing will happen from the click.
+     * There may be cases where nothing will happen and this is value is
+     * not provided, but it is guaranteed that this value is accurate
+     * when given.
+     */
+    NOTHING,
+    /**
+     * All of the items on the clicked slot are moved to the cursor.
+     */
+    PICKUP_ALL,
+    /**
+     * Some of the items on the clicked slot are moved to the cursor.
+     */
+    PICKUP_SOME,
+    /**
+     * Half of the items on the clicked slot are moved to the cursor.
+     */
+    PICKUP_HALF,
+    /**
+     * One of the items on the clicked slot are moved to the cursor.
+     */
+    PICKUP_ONE,
+    /**
+     * All of the items on the cursor are moved to the clicked slot.
+     */
+    PLACE_ALL,
+    /**
+     * Some of the items from the cursor are moved to the clicked slot
+     * (usually up to the max stack size).
+     */
+    PLACE_SOME,
+    /**
+     * A single item from the cursor is moved to the clicked slot.
+     */
+    PLACE_ONE,
+    /**
+     * The clicked item and the cursor are exchanged.
+     */
+    SWAP_WITH_CURSOR,
+    /**
+     * The entire cursor item is dropped.
+     */
+    DROP_ALL_CURSOR,
+    /**
+     * One item is dropped from the cursor.
+     */
+    DROP_ONE_CURSOR,
+    /**
+     * The entire clicked slot is dropped.
+     */
+    DROP_ALL_SLOT,
+    /**
+     * One item is dropped from the clicked slot.
+     */
+    DROP_ONE_SLOT,
+    /**
+     * The item is moved to the opposite inventory if a space is found.
+     */
+    MOVE_TO_OTHER_INVENTORY,
+    /**
+     * The clicked item is moved to the hotbar, and the item currently
+     * there is re-added to the player's inventory.
+     */
+    HOTBAR_MOVE_AND_READD,
+    /**
+     * The clicked slot and the picked hotbar slot are swapped.
+     */
+    HOTBAR_SWAP,
+    /**
+     * A max-size stack of the clicked item is put on the cursor.
+     */
+    CLONE_STACK,
+    /**
+     * The inventory is searched for the same material, and they are put
+     * on the cursor up to {@link org.bukkit.Material#getMaxStackSize()}.
+     */
+    COLLECT_TO_CURSOR,
+    /**
+     * The progress of the drag is reset. This usually means a bad packet
+     * ordering.
+     * @see InventoryDragEvent
+     */
+    RESET_DRAG_STATE,
+    /**
+     * A drag is started.
+     * @see InventoryDragEvent
+     */
+    BEGIN_DRAG,
+    /**
+     * A slot is added to the drag.
+     * @see InventoryDragEvent
+     */
+    DRAG_ADD_SLOT,
+    /**
+     * The drag is finished.
+     * @see InventoryDragEvent
+     */
+    FINISH_DRAG,
+    /**
+     * One item from the cursor is placed in each selected slot.
+     */
+    PLACE_DRAG_SINGLE,
+    /**
+     * The cursor is split evenly across all selected slots, not to
+     * exceed the Material's max stack size, with the remainder going to
+     * the cursor.
+     */
+    PLACE_DRAG_EVEN,
+    /**
+     * An unrecognized ClickType.
+     */
+    UNKNOWN,
+    ;
+}
